manage environment variable configuration using python-dotenv
