+++ .env.example
@@ -1,0 +1,6 @@
+# LLM Configuration
+LLM_ENDPOINT_URL=http://localhost:8080/v1/
+LLM_API_KEY=dummy_api_key_for_testing
+
+# System Settings (optional)
+# LOG_LEVEL=INFO
