+++ bibra/backend/__init__.py
@@ -1,4 +1,7 @@
+from bibra.backend.config import LLMConfig
 from bibra.backend.dummy import DummyBackend
 from bibra.backend.greylitlm import GreylitLMBackend
+
+__all__ = ["DummyBackend", "GreylitLMBackend", "LLMConfig"]
 
 __all__ = ["DummyBackend", "GreylitLMBackend"]
