+++ pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.11"
 dependencies = [
     "fastapi==0.135.2",
     "pydantic-ai-slim[openai]>=1.73.0",
+    "python-dotenv>=1.2.2",
     "python-multipart>=0.0.22",
     "uvicorn==0.42.0",
 ]
